+++ backend/src/main/java/com/ssafy/backend/api/service/ApplicantService.java
@@ -15,7 +15,7 @@ public interface ApplicantService {
 
     List<ApplicantState> getRecruitingApplicant(long memberNo);
 
-    int countCurrentApplicantByBoardNo(long boardNo);
+    long countCurrentApplicantByBoardNo(long boardNo);
 
     List<ApplicantState> getApplyingApplicant(long memberNo);
 }
